requirements.txt
# 旅游规划AI助手 - LangGraph 1.0版本
# Python 3.10+ 必需

# LangGraph 核心框架
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0

# LangChain 核心库
langchain-core==1.2.0
langchain-openai==1.1.3

# 数据库
aiosqlite==0.22.0

# 环境配置
python-dotenv==1.2.1

# 核心依赖
pydantic==2.12.5
pydantic_core==2.41.5
httpx==0.28.1
openai==2.11.0
tiktoken==0.12.0

# 工具库
PyYAML==6.0.3
packaging==25.0
typing_extensions==4.15.0
uuid_utils==0.12.0
xxhash==3.6.0

# 网络和序列化
requests==2.32.5
orjson==3.11.5
ormsgpack==1.12.1
jsonpatch==1.33

# 其他依赖
tenacity==9.1.2
langsmith==0.4.59
anyio==4.12.0
certifi==2025.11.12